+++ UnitTests/BraintreePayPalTests/BTPayPalClient_Tests.swift
@@ -20,7 +20,7 @@ class BTPayPalClient_Tests: XCTestCase {
         mockAPIClient.cannedResponseBody = BTJSON(value: [
             "paymentResource": ["redirectUrl": "http://fakeURL.com"]
         ])
-        payPalClient = BTPayPalClient(apiClient: mockAPIClient, universalLink: URL(string: "https://www.paypal.com")!)
+        payPalClient = BTPayPalClient(apiClient: mockAPIClient, universalLink: URL(string: "https://www.paypal.com")!, fallbackURLScheme: "paypal")
         mockWebAuthenticationSession = MockWebAuthenticationSession()
         payPalClient.webAuthenticationSession = mockWebAuthenticationSession
     }
@@ -838,6 +838,24 @@ class BTPayPalClient_Tests: XCTestCase {
     func testHandleReturnURL_whenURLIsValid_setsBTPayPalClientToNil() {
         BTPayPalClient.handleReturnURL(URL(string: "https://mycoolwebsite.com/braintree-payments/success")!)
         XCTAssertNil(BTPayPalClient.payPalClient)
+    }
+    
+    func testHandleReturnURL_withFallBack() {
+        let url = URL(string: "paypal://braintree-payments/braintreeAppSwitchPayPal/success")!
+        BTPayPalClient.payPalClient = self.payPalClient
+        XCTAssertTrue(BTPayPalClient.canHandleReturnURL(url))
+    }
+    
+    func testHandleReturnURL_withFallBack_noSuccess() {
+        let url = URL(string: "paypal://braintree-payment/braintreeAppSwitchPayPal")!
+        BTPayPalClient.payPalClient = self.payPalClient
+        XCTAssertFalse(BTPayPalClient.canHandleReturnURL(url))
+    }
+
+    func testHandleReturnURL_whenFallBack_wrongScheme() {
+        let url = URL(string: "incorrect-scheme://braintree-payment/braintreeAppSwitchPayPal/success")!
+        BTPayPalClient.payPalClient = self.payPalClient
+        XCTAssertFalse(BTPayPalClient.canHandleReturnURL(url))
     }
     
     // MARK: - App Switch - Tokenize